# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "2.0.0",
    "title": "DREAM Lab API",
    "description": "Endpoints para la API de DREAM Lab. Puede encontrar más información en [GitHub](https://github.com/CodeCraft-Solutions-DREAM-Lab).",
    "contact": {
      "name": "CodeCraft Solutions",
      "url": "https://github.com/CodeCraft-Solutions-DREAM-Lab"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Local server"
    },
    {
      "url": "https://dreamlab-api.azurewebsites.net/",
      "description": "Live server"
    }
  ],
  "paths": {
...
    "/materiales/": {
      "post": {
        "tags": [
          "Materials"
        ],
        "summary": "Obtiene los materiales disponibles dado una sala, fecha, hora de inicio, y duración",
        "description": "Obtiene los materiales disponibles dado una sala, fecha, hora de inicio, y duración",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      },
                      "cantidadDisponible": {
                        "type": "integer"
                      },
                      "image": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idSala": {
                    "type": "integer",
                    "example": 1
                  },
                  "fecha": {
                    "type": "string",
                    "format": "date",
                    "example": "2022-12-31"
                  },
                  "horaInicio": {
                    "type": "string",
                    "example": "12:00"
                  },
                  "duracion": {
                    "type": "integer",
                    "example": 2
                  }
                }
              }
            }
          }
        }
      }
    },
    "/materiales/recomendados": {
      "post": {
        "tags": [
          "Materials"
        ],
        "summary": "Obtener materiales recomendados de una experiencia",
        "description": "Obtiene los materiales disponibles y recomendados dado una sala, experiencia, fecha, hora de inicio, y duración",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      },
                      "cantidadDisponible": {
                        "type": "integer"
                      },
                      "cantidadRecomendada": {
                        "type": "integer"
                      },
                      "image": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idSala": {
                    "type": "integer",
                    "example": 1
                  },
                  "idExperiencia": {
                    "type": "integer",
                    "example": 2
                  },
                  "fecha": {
                    "type": "string",
                    "format": "date",
                    "example": "2022-12-31"
                  },
                  "horaInicio": {
                    "type": "string",
                    "example": "12:00"
                  },
                  "duracion": {
                    "type": "integer",
                    "example": 2
                  }
                }
              }
            }
          }
        }
      }
    },
    "/materiales/bySala": {
      "post": {
        "tags": [
          "Materials"
        ],
        "summary": "Obtener materiales por sala",
        "description": "Obtiene los materiales disponibles dado una sala",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      },
                      "cantidadDisponible": {
                        "type": "integer"
                      },
                      "image": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idSala": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}